userNotifications request 운동하기 직전으로 수정
