+++ WorkoutDone/WorkoutDone/Resources/AppDelegate.swift
@@ -8,26 +8,26 @@
 import UIKit
 import RealmSwift
 import NotificationCenter
-import UserNotifications
+//import UserNotifications
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
-    var userNotificationCenter = UNUserNotificationCenter.current()
+//    var userNotificationCenter = UNUserNotificationCenter.current()
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
         let realm = try! Realm()
         print(Realm.Configuration.defaultConfiguration.fileURL)
-        UNUserNotificationCenter.current().delegate = self
+//        UNUserNotificationCenter.current().delegate = self
         
-        let authrizationOptions = UNAuthorizationOptions(arrayLiteral: [.alert, .badge, .sound])
+//        let authrizationOptions = UNAuthorizationOptions(arrayLiteral: [.alert, .badge, .sound])
         
-        userNotificationCenter.requestAuthorization(options: authrizationOptions) { _, error in
-            if let error = error {
-                print(error.localizedDescription, "UNUserNotificationCenter")
-            }
-        }
+//        userNotificationCenter.requestAuthorization(options: authrizationOptions) { _, error in
+//            if let error = error {
+//                print(error.localizedDescription, "UNUserNotificationCenter")
+//            }
+//        }
         
         return true
     }
